Module sacn::error::errors

The errors used within the SacnLibrary. The ErrorKind subsection of this within the documentation contains details of all the errors.

Errors from external sources are wrapped within this error-chain.

Io errors from std::io::Error are wrapped within Io(::std::io::Error)

String errors from std::str::Utf8Error are wrapped within Str(::std::str::Utf8Error)

Uuid errors from uuid::ParseError are wrapped within Uuid(uuid::ParseError)

ParsePack related errors come within their own family wrapped inside this error to allow easy matching (can just match for SacnParsePackError rather than a specific).

SacnParsePackError(sacn_parse_pack_error::Error, sacn_parse_pack_error::ErrorKind)

Uses the error-chain crate to allow errors to allow more informative backtraces through error chaining. https://docs.rs/error-chain/0.12.2/error_chain/

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.