Function sacn::receive::check_seq_number

fn check_seq_number(
    src_sequences: &mut HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>, 
    source_limit: Option<usize>, 
    cid: Uuid, 
    sequence_number: u8, 
    universe: u16, 
    announce_timeout: bool
) -> Result<()>

Checks the given sequence number for the given universe against the given expected sequence numbers.

Returns Ok(()) if the packet is detected in-order.

Arguments

src_sequences: A mutable hashmap which relates sources identified by Uuid to another hashmap which itself relates universes to sequence numbers. The given hashmap of sequences should be for the specific packet-type being checked as different packet-types have their own sequence numbers even from the same source. source_limit: The limit on the number of sources which are allowed, None indicates no limit, if there is a limit then a SourcesExceededError may be returned. cid: The Uuid of the source that send the packet. sequence_number: The sequence number of the packet to check. universe: The universe of the packet (this is the data universe for data packets and the sync universe for synchronisation packets).

Errors

Returns an OutOfSequence error if a packet is received out of order as detected by the different between the packets sequence number and the expected sequence number as specified in ANSI E1.31-2018 Section 6.7.2 Sequence Numbering.

Return a SourcesExceededError if the cid of the source is new and would cause the number of sources to exceed the given source_limit.