Function sacn::receive::discard_lowest_priority_then_previous

pub fn discard_lowest_priority_then_previous(
    i: &DMXData, 
    n: &DMXData
) -> Result<DMXData>

The default merge action for the receiver.

This discarding of the old data is the default action for compliance as specified in ANSI E1.31-2018, Section 11.2.1.

This can be changed if required as part of the mechanism described in ANSI E1.31-2018, Section 6.2.3.4 Requirements for Merging and Arbitrating.

The first argument (i) is the existing data, n is the new data. This function is only valid if both inputs have the same universe, sync addr, start_code and the data contains at least the first value (the start code). If this doesn't hold an error will be returned. Other merge functions may allow merging different start codes or not check for them.