Function sacn::receive::remove_source_universe_seq

fn remove_source_universe_seq(
    src_sequences: &mut HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>, 
    src_cid: Uuid, 
    universe: u16
) -> Result<()>

Removes the sequence number entry from the given sequences for the given source cid and universe.

This removes the source entirely if there are no universes left.

Arguments

src_sequences: The sequence numbers for each source and universe.

src_cid: The CID for the source to remove the universe from.

universe: The universe to remove from the source.

Errors

Returns a SourceNotFound error if the given src_cid isn't in the given collection of sources/sequence-numbers.

Returns a UniverseNotFound error if the given universe isn't registered to the given source and so cannot be removed.