Struct sacn::receive::SacnReceiver

pub struct SacnReceiver {
    receiver: SacnNetworkReceiver,
    waiting_data: HashMap<u16, DMXData>,
    universes: Vec<u16>,
    discovered_sources: Vec<DiscoveredSacnSource>,
    merge_func: fn(_: &DMXData, _: &DMXData) -> Result<DMXData>,
    partially_discovered_sources: Vec<DiscoveredSacnSource>,
    source_limit: Option<usize>,
    sequences: SequenceNumbering,
    process_preview_data: bool,
    announce_source_discovery: bool,
    announce_stream_termination: bool,
    announce_timeout: bool,
}

Allows receiving dmx or other (different startcode) data using sacn.

Examples

// Example showing creation of a receiver and receiving some data, as there is no sender this receiver then handles the timeout.
use sacn::receive::SacnReceiver;
use sacn::packet::ACN_SDT_MULTICAST_PORT;
 
use std::net::{IpAddr, Ipv4Addr, SocketAddr};
use std::time::Duration;
 
const UNIVERSE1: u16 = 1;
const TIMEOUT: Option<Duration> = Some(Duration::from_secs(1)); // A timeout of None means blocking behaviour, some indicates the actual timeout.
 
let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), ACN_SDT_MULTICAST_PORT);

let mut dmx_rcv = SacnReceiver::with_ip(addr, None).unwrap();

dmx_rcv.listen_universes(&[UNIVERSE1]).unwrap();
 
match dmx_rcv.recv(TIMEOUT) {
    Err(e) => {
        // Print out the error.
        println!("{:?}", e);
    }
    Ok(p) => {
        // Print out the packet.
        println!("{:?}", p);
    }
}

Fields
receiver: SacnNetworkReceiver

The SacnNetworkReceiver used for handling communication with UDP / Network / Transport layer.

waiting_data: HashMap<u16, DMXData>

Data that hasn't been passed up yet as it is waiting e.g. due to universe synchronisation. Key is the universe. A receiver may not have more than one packet waiting per data_universe. Data_universe used as key as oppose to sync universe because multiple packets might be waiting on the same sync universe and adding data by data universe is at least as common as retrieving data by sync address because in a normal setup 1 or more bits of data wait for 1 sync.

universes: Vec<u16>

Universes that this receiver is currently listening for.

discovered_sources: Vec<DiscoveredSacnSource>

Sacn sources that have been discovered by this receiver through universe discovery packets.

merge_func: fn(_: &DMXData, _: &DMXData) -> Result<DMXData>

The merge function used by this receiver if DMXData for the same universe and synchronisation universe is received while there is already DMXData waiting for that universe and synchronisation address.

partially_discovered_sources: Vec<DiscoveredSacnSource>

Sacn sources that have been partially discovered by only some of their universes being discovered so far with more pages to go.

source_limit: Option<usize>

The limit to the number of sources for which to track sequence numbers. A new source after this limit will cause a SourcesExceededError as per ANSI E1.31-2018 Section 6.2.3.3.

sequences: SequenceNumbering

The sequence numbers being tracked by this receiver for each packet type, source and universe.

process_preview_data: bool

Flag that indicates if this receiver should process packets marked as preview data. If true then the receiver will process theses packets. Returned data contains a flag to indicate if it is preview_data which can be used by the implementer to use/discard as required.

announce_source_discovery: bool

Flag which indicates if a SourceDiscovered error should be thrown when receiving data and a source is discovered.

announce_stream_termination: bool

Flag which indicates if a StreamTerminated error should be thrown if a receiver receives a stream terminated packet.

announce_timeout: bool

Flag which indicates if an UniverseTimeout error should be thrown if it is detected that a source has timed out.

Methods

impl SacnReceiver

pub fn with_ip(
ip: SocketAddr,
source_limit: Option<usize>
) -> Result<SacnReceiver>

Creates a new SacnReceiver.

SacnReceiverInternal is used for actually receiving the sACN data but is wrapped in SacnReceiver to allow the update thread to handle timeout etc.

By default for an IPv6 address this will only receive IPv6 data but IPv4 can also be enabled by calling set_ipv6_only(false). A receiver with an IPv4 address will only receive IPv4 data.

IPv6 multicast is unsupported on Windows in Rust. This is due to the underlying library (Socket2) not providing support. Since UniverseDiscovery is primarily based around multicast to receive the UniverseDiscovery packets this mechanism is expected to have limited usage when running in an Ipv6 Windows environment. The is_multicast_enabled method can be used to see if multicast is enabled or not.

Arguments: ip: The address of the interface for this receiver to join, by default this address should use the ACN_SDT_MULTICAST_PORT as defined in ANSI E1.31-2018 Appendix A: Defined Parameters (Normative) however another address might be used in some situations. source_limit: The limit to the number of sources, past this limit a new source will cause a SourcesExceededError as per ANSI E1.31-2018 Section 6.2.3.3. A source limit of None means no limit to the number of sources.

Errors

Will return an InvalidInput error if the source_limit has a value of Some(0) which would indicate this receiver can never receive from any source.

Will return an error if the SacnReceiver fails to bind to a socket with the given ip. For more details see socket2::Socket::new().

Will return an error if the created SacnReceiver fails to listen to the E1.31_DISCOVERY_UNIVERSE. For more details see SacnReceiver::listen_universes().

pub fn set_is_multicast_enabled(&mut self, val: bool) -> Result<()>

Sets the value of the is_multicast_enabled flag to the given value.

If set to false then the receiver won't attempt to join any more multicast groups.

This method does not attempt to leave multicast groups already joined through previous listen_universe calls.

Arguments

val: The new value for the is_multicast_enabled flag.

Errors

Will return an OsOperationUnsupported error if attempting to set the flag to true in an environment that multicast isn't supported i.e. Ipv6 on Windows.

pub fn is_multicast_enabled(&self) -> bool

Returns true if multicast is enabled on this receiver and false if not. This flag is set when the receiver is created as not all environments currently support IP multicast. E.g. IPv6 Windows IP Multicast is currently unsupported.

pub fn reset_sources(&mut self)

Wipes the record of discovered and sequence number tracked sources. This is one way to handle a sources exceeded condition.

If you want to wipe data awaiting synchronisation then see (clear_all_waiting_data)[clear_all_waiting_data].

pub fn clear_all_waiting_data(&mut self)

Deletes all data currently waiting to be passed up - e.g. waiting for a synchronisation packet.

This allows clearing all data awaiting synchronisation but without forgetting sequence numbers. To wipe sequence numbers and discovered sources see (reset_sources)[reset_sources].

To clear only a specific universe of waiting data see (clear_waiting_data)[clear_waiting_data].

pub fn clear_waiting_data(&mut self, universe: u16) -> bool

Clears data (if any) waiting to be passed up for the specific universe.

Returns true if data was removed and false if there wasn't any data to remove for this universe.

Arguments

universe: The universe that the data that is waiting was sent to.

pub fn set_merge_fn(
&mut self,
func: fn(_: &DMXData, _: &DMXData) -> Result<DMXData>
) -> Result<()>

Sets the merge function to be used by this receiver.

This merge function is called if data is waiting for a universe e.g. for synchronisation and then further data for that universe with the same synchronisation address arrives.

This merge function MUST return a DmxMergeError if there is a problem merging. This error can optionally encapsulate further errors using the Error-chain system to provide a more informative backtrace.

Arguments: func: The merge function to use. Should take 2 DMXData references as arguments and return a Result.

pub fn set_ipv6_only(&mut self, val: bool) -> Result<()>

Allow only receiving on Ipv6.

pub fn listen_universes(&mut self, universes: &[u16]) -> Result<()>

Allows receiving from the given universe and starts listening to the multicast addresses which corresponds to the given universe.

Note that if the is_multicast_enabled flag is set to false then this method will only register the universe to listen to and won't attempt to join any multicast groups.

If 1 or more universes in the list are already being listened to this method will have no effect for those universes only.

Errors

Returns an ErrorKind::IllegalUniverse error if the given universe is outwith the allowed range of universes, see (is_universe_in_range)[fn.is_universe_in_range.packet].

pub fn mute_universe(&mut self, universe: u16) -> Result<()>

Stops listening to the given universe.

Errors

Returns an ErrorKind::IllegalUniverse error if the given universe is outwith the allowed range of universes, see (is_universe_in_range)[fn.is_universe_in_range.packet].

Returns UniverseNotFound if the given universe wasn't already being listened to.

pub fn set_process_preview_data(&mut self, val: bool)

Set the process_preview_data flag to the given value.

This flag indicates if this receiver should process packets marked as preview_data or should ignore them.

Argument: val: The new value of process_preview_data flag.

pub fn is_listening(&self, universe: &u16) -> bool

Checks if this receiver is currently listening to the given universe.

A receiver is 'listening' to a universe if it allows that universe to be received without filtering it out. This does not mean that the multicast address for that universe is or isn't being listened to.

Arguments: universe: The sACN universe to check

Returns: True if the universe is being listened to by this receiver, false if not.

pub fn recv(&mut self, timeout: Option<Duration>) -> Result<Vec<DMXData>>

Attempt to receive data from any of the registered universes. This is the main method for receiving data. Any data returned will be ready to act on immediately i.e. waiting e.g. for universe synchronisation is already handled.

Errors

This method will return a WouldBlock (unix) or TimedOut (windows) error if there is no data ready within the given timeout. A timeout of duration 0 will do timeout checks but otherwise will return a WouldBlock/TimedOut error without checking for data.

Will return ErrorKind::SourceDiscovered error if the announce_source_discovery flag is set and a universe discovery packet is received and a source fully discovered.

Will return a UniverseNotRegistered error if this method is called with an infinite timeout, no registered data universes and the announce_discovered_sources flag set to off. This is to protect the user from making this mistake leading to the method never being able to return.

The method may also return an error if there is an issue setting a timeout on the receiver. See SacnNetworkReceiver::set_timeout for details.

The method may also return an error if there is an issue handling the data as either a Data, Synchronisation or Discovery packet. See the SacnReceiver::handle_data_packet, SacnReceiver::handle_sync_packet and SacnReceiver::handle_universe_discovery_packet methods for details.

If the announce_timeout flag is set then the recv will return a UniverseTimeout error if a source fails to send on a universe within the timeout specified by E131_NETWORK_DATA_LOSS_TIMEOUT (ANSI E1.31-2018 Appendix A). This may not be detected immediately unless data is received for the timed-out universe from the source. If it isn't detected immediately it will be detected within an interval of E131_NETWORK_DATA_LOSS_TIMEOUT (assuming code executes in zero time).

pub fn get_announce_source_discovery(&self) -> bool

Returns the current value of the announce_source_discovery flag. See (set_announce_source_discovery)[receive::set_announce_source_discovery] for an explanation of the flag.

pub fn get_discovered_sources_no_check(&mut self) -> Vec<DiscoveredSacnSource>

Gets all discovered sources without checking if any are timed out. As the sources may be timed out get_discovered_sources is the preferred method but this is included to allow receivers to disable universe discovery source timeouts which may be useful in very high latency networks.

pub fn get_discovered_sources(&mut self) -> Vec<DiscoveredSacnSource>

Returns a list of the sources that have been discovered on the network by this receiver through the E1.31 universe discovery mechanism.

pub fn set_announce_source_discovery(&mut self, new_val: bool)

Sets the value of the announce_source_discovery flag to the given value.

By default this flag is false which indicates that when receiving data discovered sources through universe discovery won't be announced by the recv method and the receivers list of discovered universes will be updated silently. If set to true then it means that a SourceDiscovered error will be thrown whenever a source is discovered through a complete universe discovery packet.

Arguments:

new_val: The new value for the announce_source_discovery flag.

pub fn get_announce_timeout(&self) -> bool

Returns the current value of the announce_timeout flag. See (set_announce_timeout)[set_announce_timeout] for an explanation of the flag.

pub fn set_announce_timeout(&mut self, new_val: bool)

Sets the value of the announce_timeout flag to the given value.

By default this flag is false which means that if a universe for a source times out due to data not being sent then this will be updated on the receiver silently. If set to true then a UniverseTimeout error will be thrown when attempting to receive if it is detected that a source universe has timed out as per ANSI E1.31-2018 Section 6.7.1.

Arguments:

new_val: The new value for the announce_timeout flag.

pub fn get_announce_stream_termination(&self) -> bool

Returns the current value of the announce_stream_termination flag. See (set_announce_stream_termination)[set_announce_stream_termination] for an explanation of the flag.

pub fn set_announce_stream_termination(&mut self, new_val: bool)

Sets the value of the announce_stream_termination flag to the given value.

By default this flag is false. This indicates that if a source sends a stream termination packet it will be handled silently by the receiver. If set to true then a UniverseTermination error will be thrown when attempting to receive if a termination packet is received as per ANSI E1.31-2018 Section 6.2.6.

fn handle_data_packet(
&mut self,
cid: Uuid,
data_pkt: DataPacketFramingLayer
) -> Result<Option<Vec<DMXData>>>

Handles the given data packet for this DMX receiver.

Returns the universe data if successful. If the returned value is None it indicates that the data was received successfully but isn't ready to act on.

Synchronised data packets handled as per ANSI E1.31-2018 Section 6.2.4.1.

Arguments: data_pkt: The sACN data packet to handle.

Errors

Returns an OutOfSequence error if a packet is received out of order as detected by the different between the packets sequence number and the expected sequence number as specified in ANSI E1.31-2018 Section 6.7.2 Sequence Numbering.

Returns a UniversesTerminated error if a packet is received with the stream_terminated flag set indicating that the source is no longer sending on that universe and the announce_stream_termination_flag is set to true.

Will return an DmxMergeError if there is an issue merging or replacing new and existing waiting data.

fn terminate_stream<'a>(
&mut self,
src_cid: Uuid,
source_name: Cow<'a, str>,
universe: u16
)

Removes the given universe from the discovered sACN source with the given name, also stops tracking sequence numbers for that universe / sender combination.

Note this is just a record keeping operation, it doesn't actually effect the real sACN sender it just updates the record of what universes are expected on this receiver.

If the src_cid/source_name/universe isn't currently registered then this method has no effect. This is intentional as it allows calling this function multiple times without worrying about failure because it comes to the same result. E.g. when a source terminates it sends 3 termination packets but a receiver should only terminate once.

Arguments:

src_cid: The CID of the source which is terminating a universe.

source_name: The human readable name of the sACN source to remove the universe from.

universe: The sACN universe to remove.

fn store_waiting_data(&mut self, data: DMXData) -> Result<()>

Takes the given data and tries to add it to the waiting data.

Note that a receiver will only store a single packet of data per data_universe at once.

If there is waiting data for the same universe as the data then it will be merged as per the merge_func which by default keeps the highest priority data, if the data has the same priority then the newest data is kept.

Errors

Will return an DmxMergeError if there is an issue merging or replacing new and existing waiting data.

fn handle_sync_packet(
&mut self,
cid: Uuid,
sync_pkt: SynchronizationPacketFramingLayer
) -> Result<Option<Vec<DMXData>>>

Handles the given synchronisation packet for this DMX receiver.

Synchronisation packets handled as described by ANSI E1.31-2018 Section 6.2.4.1.

Returns the released / previously blocked data if successful. If the returned Vec is empty it indicates that no data was waiting.

E1.31 Synchronization Packets occur on specific universes. Upon receipt, they indicate that any data advertising that universe as its Synchronization Address must be acted upon. In an E1.31 Data Packet, a value of 0 in the Synchronization Address indicates that the universe data is not synchronized. If a receiver is presented with an E1.31 Data Packet containing a Synchronization Address of 0, it shall discard any data waiting to be processed and immediately act on that Data Packet.

If the Synchronization Address field is not 0, and the receiver is receiving an active synchronization stream for that Synchronization Address, it shall hold that E1.31 Data Packet until the arrival of the appropriate E1.31 Synchronization Packet before acting on it.

Arguments: sync_pkt: The E1.31 synchronisation part of the synchronisation packet to handle.

Errors

Returns an OutOfSequence error if a packet is received out of order as detected by the different between the packets sequence number and the expected sequence number as specified in ANSI E1.31-2018 Section 6.7.2 Sequence Numbering.

fn rtrv_waiting_data(&mut self, sync_uni: u16) -> Vec<DMXData>

Retrieves and removes the DMX data of all waiting data with a synchronisation address matching the one provided. Returns an empty Vec if there is no data waiting.

Arguments: sync_uni: The synchronisation universe of the data that should be retrieved.

fn update_discovered_srcs(&mut self, src: DiscoveredSacnSource)

Takes the given DiscoveredSacnSource and updates the record of discovered sacn sources.

This adds the new source deleting any previous source with the same name.

Arguments: src: The DiscoveredSacnSource to update the record of discovered sacn sources with.

fn handle_universe_discovery_packet(
&mut self,
discovery_pkt: UniverseDiscoveryPacketFramingLayer
) -> Option<String>

Handles the given universe discovery packet.

This universe discovery packet might be the whole thing or may be just one page of a discovery packet. This method puts the pages to produce the DiscoveredSacnSource which is stored in the receiver.

Returns the source name if a source was fully discovered or None if the source was only partially discovered.

Arguments: discovery_pkt: The universe discovery part of the universe discovery packet to handle.

fn check_waiting_data_timeouts(&mut self)

Goes through all the waiting data and removes any which has timed out as a sync-packet for it hasn't been received within the E131_NETWORK_DATA_LOSS_TIMEOUT period as specified by ANSI E1.31-2018 Section 11.1.2.

fn remove_expired_sources(&mut self)

Goes through all discovered sources and removes any that have timed out