Struct sacn::receive::SequenceNumbering

struct SequenceNumbering {
    data_sequences: HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>,
    sync_sequences: HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>,
}

Stores information about the current expected sequence numbers for each source, universe and packet type.

Also handles timeouts of sources.

Abstracts over the internal data-structures/mechanisms used allowing them be changed.

Fields
data_sequences: HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>

The sequence numbers used for data packets, keeps a reference of the last sequence number received for each universe. Sequence numbers are always in the range [0, 255] inclusive. Each type of packet is tracked differently with respect to sequence numbers as per ANSI E1.31-2018 Section 6.7.2 Sequence Numbering. The uuid refers to the source that is sending the data.

sync_sequences: HashMap<Uuid, HashMap<u16, TimedStampedSeqNo>>

The sequence numbers used for synchronisation packets, keeps a reference of the last sequence number received for each universe. Sequence numbers are always in the range [0, 255] inclusive. Each type of packet is tracked differently with respect to sequence numbers as per ANSI E1.31-2018 Section 6.7.2 Sequence Numbering. The uuid refers to the source that is sending the data.

Methods

impl SequenceNumbering

fn new() -> SequenceNumbering

Creates a new SequenceNumbering for tracking sequence numbers for the various types of packets.

This implementation uses HashMaps internally to allow O(1) checking and updating of sequence numbers.

fn clear(&mut self)

Clears the sequence number records completely removing all sources/universes for all types of packet.

fn check_timeouts(&mut self, announce_timeout: bool) -> Result<()>

Checks the timeouts for all packet types, sources and universes with sequence numbers registed. Removes any universes for which the last_recv time was at least the given timeout amount of time ago. Any sources which have no universes after this operation are also removed.

#Arguments

announce_timeout: A flag, if true it indicates than a UniverseTimeout error should be thrown if a universe times out on a source.

fn check_data_seq_number(
&mut self,
source_limit: Option<usize>,
cid: Uuid,
sequence_number: u8,
universe: u16,
announce_timeout: bool
) -> Result<()>

Checks the sequence number is correct for a data packet with the given sequence_number and universe from the given source with given cid. Uses the given source_limit to check that it isn't exceeded.

Returns Ok(()) if the packet is detected in-order.

Arguments

source_limit: The limit on the number of sources which are allowed, None indicates no limit, if there is a limit then a SourcesExceededError may be returned.

cid: The Uuid of the source that send the packet.

sequence_number: The sequence number of the packet to check.

universe: The data universe of the packet.

Errors

Returns an OutOfSequence error if a packet is received out of order as detected by the different between the packets sequence number and the expected sequence number as specified in ANSI E1.31-2018 Section 6.7.2 Sequence Numbering.

Return a SourcesExceededError if the cid of the source is new and would cause the number of sources to exceed the given source_limit.

fn check_sync_seq_number(
&mut self,
source_limit: Option<usize>,
cid: Uuid,
sequence_number: u8,
sync_uni: u16,
announce_timeout: bool
) -> Result<()>

Checks the sequence number is correct for a sync packet with the given sequence_number and universe from the given source with given cid. Uses the given source_limit to check that it isn't exceeded.

Returns Ok(()) if the packet is detected in-order.

Arguments

source_limit: The limit on the number of sources which are allowed, None indicates no limit, if there is a limit then a SourcesExceededError may be returned.

cid: The Uuid of the source that send the packet.

sequence_number: The sequence number of the packet to check.

universe: The sync universe of the packet

Errors

Returns an OutOfSequence error if a packet is received out of order as detected by the different between the packets sequence number and the expected sequence number as specified in ANSI E1.31-2018 Section 6.7.2 Sequence Numbering.

Return a SourcesExceededError if the cid of the source is new and would cause the number of sources to exceed the given source_limit.

fn remove_seq_numbers<'a>(&mut self, src_cid: Uuid, universe: u16) -> Result<()>

Removes the sequence number tracking for the given source / universe combination. This applies to both data and sync packets.

Arguments:

src_cid: The CID of the source to remove the sequence numbers of.

universe: The universe being sent by the source from which to remove the sequence numbers.