Enum sacn::sacn_parse_pack_error::sacn_parse_pack_error::ErrorKind

pub enum ErrorKind {
    Msg(String),
    ParseInvalidData(String),
    ParseInvalidPriority(String),
    ParseInvalidPage(String),
    ParseInvalidSyncAddr(String),
    ParseInvalidUniverse(String),
    ParseInvalidUniverseOrder(String),
    PackInvalidData(String),
    PackBufferInsufficient(String),
    ParseInsufficientData(String),
    ParsePduInvalidFlags(u8),
    PduInvalidLength(usize),
    PduInvalidVector(u32),
    UuidError(String),
    Utf8Error(String),
    SourceNameInvalid(String),
    __Nonexhaustive {},
}

The kind of an error.

Variants

Msg(String)

A convenient variant for String.

ParseInvalidData(String)

When parsing packet invalid data encountered.

Arguments

msg: A message providing further details (if any) as to what data was invalid.

ParseInvalidPriority(String)

Attempted to parse a priority value that is outwith the allowed range of [0, E131_MAX_PRIORITY]. As per ANSI E1.31-2018 Section 6.2.3

Arguments

msg: A message providing further details (if any) as to why the priority valid was invalid.

ParseInvalidPage(String)

Attempted to parse a page value that is invalid - e.g. the page value is higher than the last_page value.

Arguments

msg: A message providing further details (if any) as to why the page was invalid.

ParseInvalidSyncAddr(String)

Attempted to parse a sync address value that is outwith the allowed range of [0, E131_MAX_MULTICAST_UNIVERSE]. As per ANSI E1.31-2018 Section 9.1.1.

Arguments

msg: A message providing further details (if any) as to why the synchronisation address was invalid.

ParseInvalidUniverse(String)

Attempted to parse a universe value that is outwith the allowed range of [1, E131_MAX_MULTICAST_UNIVERSE]. As per ANSI E1.31-2018 Section 9.1.1.

Arguments

msg: A message providing further details (if any) as to why the universe field was invalid.

ParseInvalidUniverseOrder(String)

Attempted to parse a packet with an invalid ordering of universes. For example a discovery packet where the universes aren't correctly ordered in assending order.

Arguments

msg: A message providing further details (if any) as to why the universe ordering was invalid.

PackInvalidData(String)

When packing a packet into a buffer invalid data encountered.

Arguments

msg: A message providing further details (if any) as to why the data couldn't be packed.

PackBufferInsufficient(String)

Supplied buffer is not large enough to pack packet into.

Arguments

msg: A message providing further details (if any) as to why the pack buffer is insufficient.

ParseInsufficientData(String)

Supplied buffer does not contain enough data.

Arguments

msg: A message providing further details (if any) as to why there was insufficient data for parsing.

ParsePduInvalidFlags(u8)

Received PDU flags are invalid for parsing.

Arguments

flags: The flags that were found which are invalid.

PduInvalidLength(usize)

Received PDU length is invalid.

Arguments

len: The length provided in the Pdu which is invalid.

PduInvalidVector(u32)

Received PDU vector is invalid/unsupported by this library.

Arguments

vec: The vector parsed which is invalid / cannot be used.

UuidError(String)

Error parsing the received UUID.

Arguments

msg: A message providing further details (if any) as to why the uuid (used for CID) couldn't be parsed.

Utf8Error(String)

Error parsing received UTF8 string.

Arguments

msg: A message providing further details (if any) as to why the string couldn't be parsed.

SourceNameInvalid(String)

Source name in packet was invalid, for example due to not being null terminated.

Arguments

msg: A message providing further details (if any) as to why the source name was invalid.

__Nonexhaustive

Fields of __Nonexhaustive

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.