Module sacn::source

The source module handles generation of sACN on the network.

Structs

SacnSource
A DMX over sACN sender.

SacnSourceInternal
Internal sACN sender, this does most of the work however is encapsulated within SacnSource to allow access by the update_thread which is used to manage sending periodic universe discovery packets.

Constants

DEFAULT_POLL_PERIOD
The poll rate of the update thread. Discovery updates are sent every E131_UNIVERSE_DISCOVERY_INTERVAL so the poll rate must be lower than or equal to this.

DEFAULT_TERMINATE_START_CODE
The default startcode used to send stream termination packets when the SacnSource is closed.

SND_UPDATE_THREAD_NAME
The name of the thread which runs periodically to perform various actions such as universe discovery adverts for the source.

Functions

perform_periodic_update
Called periodically by the source update thread.

unlock_internal
Returns the locked internal SacnSourceInternal used within the SacnSource.

unlock_internal_mut
Returns the locked internal SacnSourceInternal used within the SacnSource.