Function sacn::source::perform_periodic_update

fn perform_periodic_update(
    src: &mut Arc<Mutex<SacnSourceInternal>>
) -> Result<()>

Called periodically by the source update thread.

Is responsible for sending the periodic universe discovery packets.

Arguments:

src: A reference to the SacnSourceInternal for which to send the universe discovery packet with/from.

Errors

Returns a SourceCorrupt error if the internal source mutex has been corrupted, see (unlock_internal)[unlock_internal].

Returns an error if a discovery packet cannot be sent, see (send_universe_discovery)[fn.send_universe_discovery.source].