Struct sacn::source::SacnSource

pub struct SacnSource {
    internal: Arc<Mutex<SacnSourceInternal>>,
    update_thread: Option<JoinHandle<()>>,
}

A DMX over sACN sender.

SacnSource is used for sending sACN packets over an IP network.

Examples

// Example showing creation of a source and then sending some data.
use sacn::source::SacnSource;
use sacn::packet::ACN_SDT_MULTICAST_PORT;
use std::net::{IpAddr, SocketAddr};

let local_addr: SocketAddr = SocketAddr::new(IpAddr::V4("0.0.0.0".parse().unwrap()), ACN_SDT_MULTICAST_PORT + 1);

let mut src = SacnSource::with_ip("Source", local_addr).unwrap();

let universe: u16 = 1;                        // Universe the data is to be sent on.
let sync_uni: Option<u16> = None;             // Don't want the packet to be delayed on the receiver awaiting synchronisation.
let priority: u8 = 100;                       // The priority for the sending data, must be 1-200 inclusive,  None means use default.
let dst_ip: Option<SocketAddr> = None;        // Sending the data using IP multicast so don't have a destination IP.

src.register_universe(universe).unwrap(); // Register with the source that will be sending on the given universe.

let mut data: Vec<u8> = vec![0, 0, 0, 0, 255, 255, 128, 128]; // Some arbitrary data, must have length <= 513 (including start-code).

src.send(&[universe], &data, Some(priority), dst_ip, sync_uni).unwrap(); // Actually send the data

An ANSI E1.31-2018 sACN source.

Allows sending DMX data over an IPv4 or IPv6 network using sACN.

Fields
internal: Arc<Mutex<SacnSourceInternal>>

The DMX source used for actually sending the sACN packets. Protected by a Mutex lock to allow concurrent access between user threads and the update thread below.

update_thread: Option<JoinHandle<()>>

Update thread which performs actions every DEFAULT_POLL_PERIOD such as checking if a universe discovery packet should be sent.

Methods

impl SacnSource

pub fn new_v4(name: &str) -> Result<SacnSource>

Constructs a new SacnSource with the given name, binding to an IPv4 address. This generates a new CID automatically using random values.

Errors

See (with_cid_ip)[with_cid_ip]

pub fn with_cid_v4(name: &str, cid: Uuid) -> Result<SacnSource>

Constructs a new SacnSource with the given name and specified CID binding to an IPv4 address.

Errors

See (with_cid_ip)[with_cid_ip]

pub fn new_v6(name: &str) -> Result<SacnSource>

Constructs a new SacnSource with the given name, binding to an IPv6 address. By default this will only receive IPv6 data but IPv4 can also be enabled by calling set_ipv6_only(false).

Errors

See (with_cid_ip)[with_cid_ip]

pub fn with_cid_v6(name: &str, cid: Uuid) -> Result<SacnSource>

Constructs a new SacnSource with the given name and specified CID binding to an IPv6 address.

Errors

See (with_cid_ip)[with_cid_ip]

pub fn with_ip(name: &str, ip: SocketAddr) -> Result<SacnSource>

Constructs a new SacnSource with the given name and binding to the supplied ip.

Errors

See (with_cid_ip)[with_cid_ip]

pub fn with_cid_ip(name: &str, cid: Uuid, ip: SocketAddr) -> Result<SacnSource>

Constructs a new SacnSource with the given name, cid and binding to the supplied ip.

Errors

Io: Returned if the underlying UDP socket cannot be created and bound or if the thread used for sending periodic discovery adverts fails to be created. Causes can be distinguished by looking at the error chain.

UnsupportedIpVersion: Returned if the SocketAddr is not IPv4 or IPv6.

MalformedSourceName: Returned if the given source name is longer than the maximum allowed size of E131_SOURCE_NAME_FIELD_LENGTH.

pub fn register_universes(&mut self, universes: &[u16]) -> Result<()>

Registers the given universes on this source in addition to already registered universes.

This allows sending data to those universes or using them as synchronisation addresses as well as adding them to the list of universes that appear in universe discovery packets that are sent (depending on the set_is_sending_discovery flag) periodically.

This is more efficient than repeated calls to register_universe as it means only 1 mutex unlock is required.

Arguments

universes: The sACN universes to register for usage as data universes and/or synchronisation addresses. Note that sACN universes start at 1 not 0.

Errors

IllegalUniverse: Returned if a universe is outwith the range permitted by ANSI E1.31-2018.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn register_universe(&mut self, universe: u16) -> Result<()>

Registers a single universe on this source in addition to already registered universes.

This allows sending data to those universes or using them as synchronisation addresses as well as adding them to the list of universes that appear in universe discovery packets that are sent (depending on the set_is_sending_discovery flag) periodically.

If registering multiple universes see (register_universes)[register_universes].

Arguments

universe: The sACN universe to register for usage as a data universe and/or synchronisation address. Note that sACN universes start at 1 not 0.

Errors

IllegalUniverse: Returned if the universe is outwith the range permitted by ANSI E1.31-2018.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn send(
&mut self,
universes: &[u16],
data: &[u8],
priority: Option<u8>,
dst_ip: Option<SocketAddr>,
synchronisation_addr: Option<u16>
) -> Result<()>

Sends the given data to the given universes with the given priority, synchronisation address (universe) and destination ip.

Arguments

universe: The sACN universes that the data should be set on, the data will be split over these universes with each UNIVERSE_CHANNEL_CAPACITY sized chunk sent to the next universe.

data: The data that should be sent, must have a length greater than 0.

priority: The E131 priority that the data should be sent with, must be less than E131_MAX_PRIORITY (const.E131_MAX_PRIORITY.packet), if a value of None is provided then the default of E131_DEFAULT_PRIORITY (const.E131_DEFAULT_PRIORITY.packet) is used.

dst_ip: The destination IP, can be Ipv4 or Ipv6, None if should be sent using ip multicast.

sync_address: The address to use for synchronisation, must be a valid universe, None indicates no synchronisation. If synchronisation is required a reasonable default address to use is the first universe that this data is being sent to.

As per ANSI E1.31-2018 Section 6.6.1 this method shouldn't be called at a higher refresher rate than specified in ANSI E1.11 [DMX] unless configured by the user to do so in an environment which doesn't contain any E1.31 to DMX512-A converters.

Note as per ANSI-E1.31-2018 Appendix B.1 it is recommended to have a small delay before sending the follow up sync packet.

Errors

SenderAlreadyTerminated: Returned if this method is called on an SacnReceiverInternal that has already terminated.

InvalidInput: Returned if the data array has length 0 or if an insufficient number of universes for the given data are provided (each universe takes 513 bytes of data).

InvalidPriority: Returned if the priority is greater than the allowed maximum priority of E131_MAX_PRIORITY.

IllegalUniverse: Returned if the universe is outwith the allowed range as specified by ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

ExceedUniverseCapacity: Returned if the data has a length greater than the maximum allowed within a universe (packet::UNIVERSE_CHANNEL_CAPACITY).

Io: Returned if the data fails to be sent on the socket, see send_to(fn.send_to.Socket).

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn send_sync_packet(
&mut self,
universe: u16,
dst_ip: Option<SocketAddr>
) -> Result<()>

Sends a synchronisation packet to trigger the sending of packets waiting to be sent together.

A common pattern would be to use the send method to send data to all the universes that should be synchronised using a chosen synchronisation universe then wait for a small time as per the recommendation in ANSI-E1.31-2018 Appendix B.1 and then send a synchronisation packet with the address of the synchronisation universe chosen to trigger the packets.

Arguments

universe: The universe of this synchronisation packet. dst_ip: The destination IP address for this packet or None if it should be sent using multicast.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range of sACN universes as defined in ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

Io: Returned if the packet fails to be sent using the underlying network socket.

SacnParsePackError: Returned if the sync packet fails to be packed.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn terminate_stream(&mut self, universe: u16, start_code: u8) -> Result<()>

Terminates sending on the given universe.

Errors:

IllegalUniverse: Returned if the universe is outwith the allowed range of sACN universes as defined in ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on this source.

Io: Returned if the termination packets fail to be sent on the socket.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn cid(&self) -> Result<Uuid>

Returns the ACN CID device identifier of the SacnSourceInternal.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_cid(&mut self, cid: Uuid) -> Result<()>

Sets the ACN CID device identifier.

Arguments

cid: The new CID identifier for this source. It is left to the user to ensure that this is always unique within the network the source is in.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn name(&self) -> Result<String>

Returns the ACN source name.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_name(&mut self, name: &str) -> Result<()>

Sets ACN source name.

Argument

name: The new name for the source, it is left to the user to ensure this is unique within the sACN network.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

MalformedSourceName: Returned to indicate that the given source name is longer than the maximum allowed as per E131_SOURCE_NAME_FIELD_LENGTH.

pub fn preview_mode(&self) -> Result<bool>

Returns true if SacnSourceInternal is in preview mode, false if not.

For details of preview_mode see (set_preview_mode)[set_preview_mode].

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_preview_mode(&mut self, preview_mode: bool) -> Result<()>

Sets the value of the Preview_Data flag in packets from this SacnSource.

Arguments

preview_mode: If true then all data packets from this SacnSource will have the Preview_Data flag set to true indicating that the data is not for live output. If false then the flag will be set to false.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_is_sending_discovery(&mut self, val: bool)

Sets the is_sending_discovery flag to the given value.

Arguments

val: The new value for the is_sending_discovery flag, if true then source will send periodic universe discovery packets and if false it won't.

pub fn multicast_ttl(&self) -> Result<u32>

Returns the multicast time to live of the socket.

pub fn set_multicast_ttl(&mut self, multicast_ttl: u32) -> Result<()>

Sets the multicast time to live.

Arguments

multicast_ttl: The new time to live value for network packets sent using multicast.

Errors

Io: Returned if the multicast TTL fails to be set on the underlying socket.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn ttl(&self) -> Result<u32>

Returns the current Time To Live for unicast packets send by this source.

Errors

Io: Returned if the TTL cannot be retrieved from the underlying socket.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_ttl(&mut self, ttl: u32) -> Result<()>

Sets the Time To Live for packets sent by this source.

Arguments

ttl: The new time to live value for new packets.

Errors

Io: Returned if the TTL value cannot be changed.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn set_multicast_loop_v4(&mut self, multicast_loop: bool) -> Result<()>

Sets if multicast loop is enabled.

Arguments:

multicast_loop: If true then multicast loop is enabled, if false it is not.

Errors

Io: Returned if the set_multicast_loop option fails to be set on the socket.

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn multicast_loop(&self) -> Result<bool>

Returns true if multicast loop is enabled, false if not.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.

pub fn universes(&self) -> Result<Vec<u16>>

Returns the universes currently registered on this source.

Errors

SourceCorrupt: Returned if the Mutex used to control access to the internal sender is poisoned by a thread encountering a panic while accessing causing the source to be left in a potentially inconsistent state.