Struct sacn::source::SacnSourceInternal

struct SacnSourceInternal {
    socket: Socket,
    addr: SocketAddr,
    cid: Uuid,
    name: String,
    preview_data: bool,
    data_sequences: RefCell<HashMap<u16, u8>>,
    sync_sequences: RefCell<HashMap<u16, u8>>,
    universes: Vec<u16>,
    running: bool,
    last_discovery_advert_timestamp: Instant,
    is_sending_discovery: bool,
}

Internal sACN sender, this does most of the work however is encapsulated within SacnSource to allow access by the update_thread which is used to manage sending periodic universe discovery packets.

Fields
socket: Socket

Underlying UDP socket used for sending sACN packets on the network.

addr: SocketAddr

The address of this SacnSourceInternal on the network.

cid: Uuid

The unique ID of this SacnSourceInternal. It is the job of the user of the library to ensure that the cid is given on creation of the SacnSourceInternal is unique.

name: String

The human readable name of this source.

preview_data: bool

Flag which is included in sACN packets to indicate that the data shouldn't be used for live output (ie. on actual lighting fixtures). A receiver may or may not be compliant with this so it should not be relied upon in an untested environment.

data_sequences: RefCell<HashMap<u16, u8>>

The sequence numbers used for data packets, keeps a reference of the next sequence number to use for each universe. Sequence numbers are always in the range [0, 255].

sync_sequences: RefCell<HashMap<u16, u8>>

The sequence numbers used for sync packets, keeps a reference of the next sequence number to use for each universe. Sequence numbers are always in the range [0, 255].

universes: Vec<u16>

A list of the universes registered to send by this source, used for universe discovery. Always sorted with lowest universe first to allow quicker usage. This may never contain duplicate universe values.

running: bool

Flag that indicates if the SacnSourceInternal is running (the update thread should be triggering periodic discovery packets).

last_discovery_advert_timestamp: Instant

The time that the last universe discovery advert was send.

is_sending_discovery: bool

Flag that is set to True to indicate that the source is sending periodic universe discovery packets.

Methods

impl SacnSourceInternal

fn with_cid_ip(
name: &str,
cid: Uuid,
ip: SocketAddr
) -> Result<SacnSourceInternal>

Constructs a new SacnSourceInternal with DMX START code set to 0 with specified CID and binding IP address.

By default for an IPv6 address this will only receive IPv6 data but IPv4 can also be enabled by calling set_ipv6_only(false). By default the TTL for ipv4 packets is 1 to keep them within the local network.

Arguments:

name: The human readable name for this sacn source. cid: The UUID for this source. ip: The address that this source should bind to.

Errors

Io: Returned if the underlying socket cannot be created or the IP cannot be bound to the underlying socket. See (UdpBuilder::new_v4)[fn.new_v4.UdpBuilder], (UdpBuilder::new_v6)[fn.new_v6.UdpBuilder] and (Socket::bind)[fn.bind.Socket2].

UnsupportedIpVersion: Returned if the SockAddr is not IPv4 or IPv6.

fn set_is_sending_discovery(&mut self, val: bool)

Sets the is_sending_discovery flag to the given value.

If is_sending_discovery is set to false then no discovery adverts for this source will be sent otherwise (and by default) they will be sent every UNIVERSE_DISCOVERY_INTERVAL.

Arguments:

val: The new value of the is_sending_discovery flag.

fn register_universes(&mut self, universes: &[u16]) -> Result<()>

Registers the given array of universes with this source.

Any universes already registered won't be re-registered and will have no effect.

Arguments:

universes: The sACN universe to register. Note that sACN universes start at 1 not 0.

Errors

See register_universe(fn.register_universe.source) for more details.

fn register_universe(&mut self, universe: u16) -> Result<()>

Registers the given universe for sending with this source.

If a universe is already registered then this method has no effect.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range, see (is_universe_in_range)[fn.is_universe_in_range.packet].

fn deregister_universe(&mut self, universe: u16) -> Result<()>

De-registers the given universe for sending with this source.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range, see (is_universe_in_range)[fn.is_universe_in_range.packet].

UniverseNotFound: Returned if the given universe was never registered originally.

fn universe_allowed(&self, u: &u16) -> Result<()>

Checks if the given universe is a valid universe to send on (within allowed range) and that it is registered with this SacnSourceInternal.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range, see (is_universe_in_range)[fn.is_universe_in_range.packet].

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

fn send(
&self,
universes: &[u16],
data: &[u8],
priority: Option<u8>,
dst_ip: Option<SocketAddr>,
synchronisation_addr: Option<u16>
) -> Result<()>

Sends the given data to the given universes with the given priority, synchronisation address (universe) and destination ip.

Arguments

universe: The sACN universes that the data should be set on, the data will be split over these universes with each UNIVERSE_CHANNEL_CAPACITY sized chunk sent to the next universe.

data: The data that should be sent, must have a length greater than 0.

priority: The E131 priority that the data should be sent with, must be less than E131_MAX_PRIORITY (const.E131_MAX_PRIORITY.packet), if a value of None is provided then the default of E131_DEFAULT_PRIORITY (const.E131_DEFAULT_PRIORITY.packet) is used.

dst_ip: The destination IP, can be Ipv4 or Ipv6, None if should be sent using ip multicast.

sync_address: The address to use for synchronisation, must be a valid universe, None indicates no synchronisation. If synchronisation is required a reasonable default address to use is the first universe that this data is being sent to.

As per ANSI E1.31-2018 Section 6.6.1 this method shouldn't be called at a higher refresher rate than specified in ANSI E1.11 [DMX] unless configured by the user to do so in an environment which doesn't contain any E1.31 to DMX512-A converters.

Note as per ANSI-E1.31-2018 Appendix B.1 it is recommended to have a small delay before sending the follow up sync packet.

Errors

SenderAlreadyTerminated: Returned if this method is called on an SacnReceiverInternal that has already terminated.

InvalidInput: Returned if the data array has length 0 or if an insufficient number of universes for the given data are provided (each universe takes 513 bytes of data).

InvalidPriority: Returned if the priority is greater than the allowed maximum priority of E131_MAX_PRIORITY.

IllegalUniverse: Returned if the universe is outwith the allowed range as specified by ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

ExceedUniverseCapacity: Returned if the data has a length greater than the maximum allowed within a universe (packet::UNIVERSE_CHANNEL_CAPACITY).

Io: Returned if the data fails to be sent on the socket, see send_to(fn.send_to.Socket).

fn send_universe(
&self,
universe: u16,
data: &[u8],
priority: u8,
dst_ip: &Option<SocketAddr>,
sync_address: u16
) -> Result<()>

Sends the given data to the given universe with the given priority, synchronisation address (universe) and destination ip.

Arguments

universe: The sACN universe that the data should be set on.

data: The data that should be sent, must be less than or equal in length to UNIVERSE_CHANNEL_CAPACITY(const.UNIVERSE_CHANNEL_CAPACITY.packet).

priority: The E131 priority that the data should be sent with, must be less than E131_MAX_PRIORITY (const.E131_MAX_PRIORITY.packet), default E131_DEFAULT_PRIORITY.

dst_ip: The destination IP, can be Ipv4 or Ipv6, None if should be sent using ip multicast.

sync_address: The address to use for synchronisation, must be a valid universe, 0 indicates no synchronisation.

Errors

InvalidInput: Returned if the priority is greater than the allowed maximum priority of E131_MAX_PRIORITY.

ExceedUniverseCapacity: Returned if the data has a length greater than the maximum allowed within a universe.

IllegalUniverse: Returned if the given universe is outwith the allowed range of universes, see (universe_to_ipv4_multicast_addr)[fn.universe_to_ipv4_multicast_addr.packet] and (universe_to_ipv6_multicast_addr)[fn.universe_to_ipv6_multicast_addr.packet].

Io: Returned if the data fails to be sent on the socket, see send_to(fn.send_to.Socket).

fn send_sync_packet(
&self,
universe: u16,
dst_ip: Option<SocketAddr>
) -> Result<()>

Sends a synchronisation packet to trigger the sending of packets waiting to be sent together.

A common pattern would be to use the send method to send data to all the universes that should be synchronised using a chosen synchronisation universe then wait for a small time as per the recommendation in ANSI-E1.31-2018 Appendix B.1 and then send a synchronisation packet with the address of the synchronisation universe chosen to trigger the packets.

Arguments

universe: The universe of this synchronisation packet. dst_ip: The destination IP address for this packet or None if it should be sent using multicast.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range of sACN universes as defined in ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

Io: Returned if the packet fails to be sent using the underlying network socket.

SacnParsePackError: Returned if the sync packet fails to be packed.

fn send_terminate_stream_pkt(
&self,
universe: u16,
dst_ip: Option<SocketAddr>,
start_code: u8
) -> Result<()>

Sends a stream termination packet for the given universe.

In normal usage this method would be called three times to send three packets for termination as per ANSI E1.31-2018 Section 6.2.6, Stream_Terminated: Bit 6.

Arguments

universe: The universe of this synchronisation packet. dst_ip: The destination IP address for this packet or None if it should be sent using multicast.

Errors

IllegalUniverse: Returned if the universe is outwith the allowed range of sACN universes as defined in ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on the given SacnSourceInternal.

Io: Returned if the termination packets fail to be sent on the underlying socket.

fn terminate_stream(&mut self, universe: u16, start_code: u8) -> Result<()>

Terminates a universe stream.

Terminates a stream to the specified universe by sending packets with the Stream_Terminated flag set to 1. Number of packets sent as per section 6.2.6 , Stream_Terminated: Bit 6 of ANSI E1.31-2018.

Arguments: universe: The universe that is being terminated. start_code: used for the first byte of the otherwise empty data payload to indicate the start_code of the data.

Errors:

IllegalUniverse: Returned if the universe is outwith the allowed range of sACN universes as defined in ANSI E1.31-2018 Section 6.2.7.

UniverseNotRegistered: Returned if the universe is not registered on this source.

Io: Returned if the termination packets fail to be sent on the socket.

fn terminate(&mut self, start_code: u8) -> Result<()>

Terminates the DMX source.

This includes terminating each registered universe with the start_code given.

Arguments: start_code: used for the first byte of the otherwise empty data payload to indicate the start_code of the data.

Errors:

Io: Returned if the termination packets fail to be sent on the underlying socket.

fn send_universe_discovery(&self) -> Result<()>

Sends a universe discovery packet advertising the universes that this source is registered to send.

This packet may be broken down into multiple pages internally resulting in multiple UDP packets.

Errors

See (send_universe_discovery_detailed)[fn.send_universe_discovery_detailed.source].

fn send_universe_discovery_detailed(
&self,
page: u8,
last_page: u8,
universes: &[u16]
) -> Result<()>

Sends a page of a universe discovery packet.

There may be 1 or more pages for each full universe discovery packet with each page sent separately.

Arguments

page: The page number of this universe discovery page.

last_page: The last page that is expected as part of this universe discovery packet.

universes: The universes to include on the page.

Errors

Io: Returned if the discovery packet fails to be sent on the socket.

SacnParsePackError: Returned if the discovery packet cannot be packed to send.

fn cid(&self) -> &Uuid

Returns the ACN CID device identifier of the SacnSourceInternal.

fn set_cid(&mut self, cid: Uuid)

Sets the ACN CID device identifier.

Arguments

cid: The new CID identifier for this source. It is left to the user to ensure that this is always unique within the network the source is in.

fn name(&self) -> &str

Returns the ACN source name.

fn set_name(&mut self, name: &str) -> Result<()>

Sets ACN source name.

Argument

name: The new name for the source, it is left to the user to ensure this is unique within the sACN network.

Errors

MalformedSourceName: Returned to indicate that the given source name is longer than the maximum allowed as per E131_SOURCE_NAME_FIELD_LENGTH.

fn preview_mode(&self) -> bool

Returns if SacnSourceInternal is in preview mode.

fn set_preview_mode(&mut self, preview_mode: bool)

Sets the value of the Preview_Data flag in packets from this SacnSourceInternal.

Arguments

preview_mode: If true then all data packets from this SacnSourceInternal will have the Preview_Data flag set to true indicating that the data is not for live output. If false then the flag will be set to false.

fn set_multicast_ttl(&self, multicast_ttl: u32) -> Result<()>

Sets the multicast time to live.

Arguments

multicast_ttl: The new time to live value for network packets sent using multicast.

Errors

Io: Returned if the multicast TTL fails to be set on the underlying socket.

fn ttl(&self) -> Result<u32>

Returns the current Time To Live for unicast packets send by this source.

Errors

Io: Returned if the TTL cannot be retrieved from the underlying socket.

fn set_ttl(&mut self, ttl: u32) -> Result<()>

Sets the Time To Live for unicast packets sent by this source.

Arguments

ttl: The new time to live value for network packets sent by the source.

Errors

Io: Returned if the TTL fails to be set on the underlying socket.

fn multicast_ttl(&self) -> Result<u32>

Returns the multicast time to live of the socket.

fn set_multicast_loop_v4(&self, multicast_loop: bool) -> Result<()>

Sets if multicast loop is enabled.

Arguments:

multicast_loop: If true then multicast loop is enabled, if false it is not.

Errors

Io: Returned if the set_multicast_loop option fails to be set on the socket.

fn multicast_loop(&self) -> Result<bool>

Returns true if multicast loop is enabled, false if not.

pub fn universes(&self) -> Vec<u16>

Returns the universes currently registered on this source.